requirements.txt
streamlit>=1.24.0
google-cloud-storage>=2.8.0
google-auth>=2.17.3
pandas>=1.5.3
sendgrid>=6.10.0
email-validator>=2.0.0
python-multipart>=0.0.6 